+++ src/demo/Multip.example.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { multip } from "../components";
+
+function MultipExample() {
+  return (
+    <code>
+      multip(1, 2, 3, 4, 5) = {multip(1, 2, 3, 4, 5)}
+    </code>
+  )
+}
+
+export default MultipExample;