+++ src/demo/Sum.example.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { sum } from "../components";
+
+function SumExample() {
+  return (
+    <code>
+      sum(1, 2, 3, 4, 5) = {sum(1, 2, 3, 4, 5)}
+    </code>
+  )
+}
+
+export default SumExample;