+++ src/demo/common/utils.js
@@ -1,0 +1,7 @@
+const getUrl = path => 'https://github.com/bonnv79/react-library-basic';
+
+const getSourceUrl = path => `${getUrl()}/tree/master/src/demo/${path}`;
+
+const getSrcUrl = path => `${getUrl()}/tree/master/${path}`;
+
+export { getUrl, getSourceUrl, getSrcUrl };