+++ src/demo/index.js
@@ -1,13 +1,20 @@
 import React from "react";
-import Component from "../components";
-import logo from './logo.svg';
+import Home from "./common/Home";
+import MultipExample from "./Multip.example";
+import SumExample from "./Sum.example";
+import pack from '../../package.json';
+import { getSourceUrl } from "./common/utils";
 
 function Demo() {
-
   return (
-    <div>
-      <Component title="Demo" logo={logo} />
-    </div>
+    <Home title="React" version={`v${pack && pack.version ? pack.version : ''}`}>
+      <div key="Sum" link={getSourceUrl('Sum.example.js')}>
+        <SumExample />
+      </div>
+      <div key="Multip" link={getSourceUrl('Multip.example.js')}>
+        <MultipExample />
+      </div>
+    </Home>
   )
 }